Added the rest of the economy buildings - time to optimize!
We probably want to create a Tab class owned by the UI which contains
the buttons for each type of building (so the Economy tab has the mine,
farm, etc..)
We also need a better way to handle structure creation in the
CityManager. I think having a List<EconomyBuilding>,
List<MilitaryBuilding>, and List<Defenses> may work better to reduce
code, but then we need a way to encode which resources are produced when
the building updates.