Implemented switching between tabs, hacked the CityManager to stop the game from crashing
We need a good way to handle what type of building is being placed.
Maybe do an initial switch statement for instanceOf ResourceBuilding and
instanceOf military building which then calls getResourceBuilding() or
getMilitaryBuilding()

we also need classes for each of the military buildings as well as for
the mana pumps.