Added Enemy and Buildable tiles and test map
The Buildable tile is the area that the players base will initially be
in, and the Enemy tile is the area where the enemy AI will be able to
place buildings.
I can see at least four more tiletypes we may want to add for some
initial AI: HQ, EnemyHQ, Spawn, and EnemySpawn.