Stopped AI from cheating
The AI no longer cheats to build. The EnemyCityHandler uses the
CostCheck class to determine which building it should build. It focuses
military as long as it doesn't have more than 2x as many military vs
resource buildings.

Also I opened the project in the new Eclipse, so there are a lot of
binary file changes.